requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cramjam==2.8.3
cycler==0.12.1
defusedxml==0.7.1
Django==4.2.8
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.15.0
elasticsearch==8.15.1
elasticsearch-dsl==8.15.2
fastparquet==2024.5.0
fonttools==4.47.0
fsspec==2024.9.0
httpie==3.2.2
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
kiwisolver==1.4.5
lxml==5.0.0
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
MLB-StatsAPI==0.1.0
multidict==6.0.4
multipledispatch==1.0.0
natsort==8.4.0
nba_api==1.4.1
nfl_data_py==0.3.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pandas-flavor==0.6.0
pillow==10.2.0
polars==0.18.15
pyarrow==14.0.2
Pygments==2.17.2
pyjanitor==0.26.0
PyJWT==2.8.0
pyparsing==3.1.1
pyreadr==0.5.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
scipy==1.11.4
six==1.16.0
soupsieve==2.5
sportsdataverse==0.0.39
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
xarray==2023.12.0
xgboost==2.0.3
zipp==3.17.0
